Created CreateJSON Tool for Diff and Tool Compability
This may come as a suprise addition.

However, to support the needs for:
1. Data versioning.
2. Existing tool compatbility
3. User/Newb friendliness
4. Data versioning agnostic to database/provider/user

It was nessecary to create an intermediate, human readable and diffable
version of the DBC files to be used in git/repositories. This lets
anyone use DBC tools to modify DBC data without requring EVERYONE to use
the same tools or even a database.

While I recommend you use a database and deal with SQL migrations this
is important for supporting large teams and existing tools.